+++ lib/models/request_mock.js
@@ -3,7 +3,7 @@
 const EventEmitter = require('events');
 const URL = require('url').URL;
 
-const RequestFilter = require('../request_filter');
+const RequestFilter = require('../modules/requests/request_filter');
 const {FatalError, AssertionError, TimeoutError} = require('../errors');
 const utils = require('../utils');
