+++ lib/modules/assertions/browser_assertions.js
@@ -5,7 +5,7 @@
 const utils = require('../../utils');
 const elementsAssertionUtils = require('../../utils/assert_elements');
 const assertUtils = require('../../utils/assert_utils');
-const RequestAssertionsFilter = require('./request_assertions_filter');
+const RequestAssertionsFilter = require('../requests/request_assertions_filter');
 const BrowserNotAssertions = require('./browser_not_assertions');
 const {QueryError} = require('../../errors');
