+++ lib/modules/requests/request_mocker.js
@@ -1,8 +1,8 @@
 "use strict";
 
 const URL = require('url').URL;
-const utils = require('./utils');
-const RequestMock = require('./models/request_mock');
+const utils = require('../../utils');
+const RequestMock = require('../../models/request_mock');
 
 
 module.exports = class RequestMocker {
